+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog for emr-launch
 0.5.0 (unreleased)
 ------------------
 
+- decouple emr_profile and cluster_configuration
+
 - rehydrate EMRLaunchFunction from stored function
 
 - TerminateCluster task waits for actual termination